validacion del token para actualizacion y eliminacion de tareas, corregir la consulta de una sola tarea

+++ routes/task.route.js
@@ -8,19 +8,17 @@ const { task } = require("../db/schema");
 
 //consultando todasd las tareas del usuario
 router.get("/:userid", async (req, res) => {
-  try{
-  const token= jwt.verify(req.body.token, process.env.SECRET)
-  if(token){
+  try {
+    const token = jwt.verify(req.body.token, process.env.SECRET);
+    if (token) {
       //traer las tareas
-  const homeWork = await task.find({ user: req.params.userid });
-  res.send(homeWork);
-  }
-  else{
-    res.send({message: "invalid token"})
-  }
-  }
-  catch(err){
-    res.send(err)
+      const homeWork = await task.find({ user: req.params.userid });
+      res.send(homeWork);
+    } else {
+      res.send({ message: "invalid token" });
+    }
+  } catch (err) {
+    res.send(err);
     console.log(err);
   }
 });
@@ -28,15 +26,24 @@ router.get("/:userid", async (req, res) => {
 //consultando una tarea especifica del usuario
 router.get("/:userid/:id", async (req, res) => {
   const id = req.params.id;
-  //traer la tarea especifica
-  const homeWork = await task.findOne({ user: req.params.userid, id: id });
-  //comprobando si existe la tarea
-  if (homeWork) {
-    res.send(homeWork);
-  } else {
-    res.send({ message: "task does not exist" });
+  try {
+    const token = jwt.verify(req.body.token, process.env.SECRET);
+    if (token) {
+      //traer la tarea especifica
+      const homeWork = await task.findOne({ user: req.params.userid, id: id });
+      //comprobando si existe la tarea
+      if (homeWork) {
+        res.send(homeWork);
+      } else {
+        res.send({ message: "task does not exist" });
+      }
+      // console.log("la tarea");
+    } else {
+      res.send({ message: "invalid token" });
+    }
+  } catch (err) {
+    console.log(err);
   }
-  // console.log("la tarea");
 });
 
 //creacion de una tarea
@@ -69,22 +76,37 @@ router.put("/:userid/:id", async (req, res) => {
   //creando la tarea actualizada
   const TASK = { title, description, done };
   const user = req.params.userid;
-  //encontrar la tarea especifica y actualizarla
-  await task.findOneAndUpdate({ user: user, id: req.params.id }, TASK);
-  console.log("todo bien");
-  res.send({ status: "ok" });
+  try {
+    const token = jwt.verify(req.body.token, process.env.SECRET);
+    if (token) {
+      //encontrar la tarea especifica y actualizarla
+      await task.findOneAndUpdate({ user: user, id: req.params.id }, TASK);
+      console.log("todo bien");
+      res.send({ status: "ok" });
+    } else {
+      res.send({ message: "invalid token" });
+    }
+  } catch (err) {
+    console.log(err);
+  }
 });
 
 //elimincacion de una tarea
 router.delete("/:userid/:id", async (req, res) => {
   const user = req.params.userid;
-  //encontrando la tarea especifica y borrandola
-  await task.findOneAndDelete({ user: user, id: req.params.id });
-  res.send({ status: "ok" });
-  console.log("oh shit bro");
+  try {
+    const token = jwt.verify(req.body.token, process.env.SECRET);
+    if (token) {
+      //encontrando la tarea especifica y borrandola
+      await task.findOneAndDelete({ user: user, id: req.params.id });
+      res.send({ status: "ok" });
+      console.log("oh shit bro");
+    } else {
+      res.send({ message: "invalid token" });
+    }
+  } catch (err) {
+    console.log(err);
+  }
 });
 
 module.exports = router;
-
-//eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjYzMzM1ZThkNjFkMWRiODU0Y2U0ODllOCIsImVtYWlsIjoiYWxlQGdtYWlsLmNvbSIsImlhdCI6MTY2NDM5ODU0MH0.CgRKPDTwDMXwc46CmakbrW5kVAdcLSc5-Ga-XLhq0e0
-//
